refactor: split file-browser into section files (#259)
Break the 1245-line file-browser.tsx into components/dashboard/file-browser/:
container (FileBrowser), FileRow, FileCard, FileActions + hook, the
BatchHeader trio, SelectableIcon/StatusDot, and a pure status.ts with
unit tests covering the deriveStatus lockstep invariant. Import path
is unchanged via index.ts.
