+++ apps/web/components/dashboard/file-browser/FileBrowser.tsx
@@ -14,262 +14,36 @@ import {
     AlertDialogTrigger,
 } from '@/components/ui/alert-dialog';
 import {
-    DropdownMenu,
-    DropdownMenuContent,
-    DropdownMenuItem,
-    DropdownMenuPositioner,
-    DropdownMenuSeparator,
-    DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu';
-import {
     Table,
     TableBody,
-    TableCell,
     TableHead,
     TableHeader,
     TableRow,
 } from '@/components/ui/table';
-import { Card, CardContent } from '@/components/ui/card';
+import { Card } from '@/components/ui/card';
 import {
     Search,
     LayoutGrid,
     LayoutList,
-    MoreHorizontal,
-    Download,
     Trash2,
-    Clock,
-    FileIcon,
-    FileText,
-    FileImage,
-    FileVideo,
-    FileAudio,
-    FileArchive,
-    FileCode,
-    ChevronRight,
     RotateCw,
     Loader2,
     X,
     Snowflake,
 } from 'lucide-react';
 import { cn } from '@/lib/cn';
-import { formatBytes, formatDate, formatDownloadWindow } from '@/lib/format';
 import { useTRPC } from '@/lib/trpc/client';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { useMutation, useQuery } from '@tanstack/react-query';
 import { toast } from 'sonner';
 import { useDebouncedValue } from '@/lib/hooks/useDebouncedValue';
 import { useInvalidateFileList } from '@/lib/hooks/useInvalidateFileList';
 import { RetrieveDialog } from '@/components/dashboard/RetrieveDialog';
-import type { FileBatchGroup, FileWithRetrieval } from '@nexus/db/repo/files';
-
-type DerivedStatus = 'archived' | 'retrieving' | 'available';
+import { deriveStatus } from './status';
+import { BatchHeader, BatchHeaderRow } from './BatchHeader';
+import { FileRow } from './FileRow';
+import { FileCard } from './FileCard';
 
 const SEARCH_DEBOUNCE_MS = 300;
-
-// Keep in lockstep with countStatusesByUser in
-// packages/db/src/repositories/files.ts — the library-wide stats bar bucket
-// counts must match the per-row status dots derived here.
-function deriveStatus(file: FileWithRetrieval): DerivedStatus {
-    // The active retrieval wins, uniformly across tiers (#259): a ready row
-    // is downloadable within its window, a queued/restoring row is in
-    // flight. Without one the file sits archived — never Download, since
-    // getDownloadUrl can't serve without a ready retrieval (#256).
-    if (file.activeRetrieval) {
-        return file.activeRetrieval.status === 'ready'
-            ? 'available'
-            : 'retrieving';
-    }
-    if (file.status === 'restoring') return 'retrieving';
-    return 'archived';
-}
-
-function getDownloadWindowLabel(file: FileWithRetrieval): string | null {
-    const retrieval = file.activeRetrieval;
-    if (!retrieval) return null;
-    return formatDownloadWindow(retrieval.status, retrieval.expiresAt);
-}
-
-function getFileExtension(name: string): string {
-    const parts = name.split('.');
-    return parts.length > 1 ? parts.pop()!.toLowerCase() : '';
-}
-
-function getFileTypeInfo(name: string): {
-    icon: typeof FileIcon;
-    colorClass: string;
-} {
-    const ext = getFileExtension(name);
-
-    const imageExts = [
-        'jpg',
-        'jpeg',
-        'png',
-        'gif',
-        'svg',
-        'webp',
-        'bmp',
-        'ico',
-    ];
-    const videoExts = ['mp4', 'mov', 'avi', 'mkv', 'webm', 'flv'];
-    const audioExts = ['mp3', 'wav', 'flac', 'aac', 'ogg', 'm4a'];
-    const archiveExts = ['zip', 'tar', 'gz', 'rar', '7z', 'bz2'];
-    const codeExts = [
-        'js',
-        'ts',
-        'tsx',
-        'jsx',
-        'py',
-        'rb',
-        'go',
-        'rs',
-        'java',
-        'c',
-        'cpp',
-        'h',
-        'css',
-        'html',
-        'json',
-        'yaml',
-        'yml',
-        'xml',
-        'sql',
-        'sh',
-    ];
-    const docExts = [
-        'pdf',
-        'doc',
-        'docx',
-        'txt',
-        'md',
-        'rtf',
-        'xls',
-        'xlsx',
-        'csv',
-        'ppt',
-        'pptx',
-    ];
-
-    if (imageExts.includes(ext))
-        return { icon: FileImage, colorClass: 'text-rose-500 bg-rose-500/10' };
-    if (videoExts.includes(ext))
-        return {
-            icon: FileVideo,
-            colorClass: 'text-purple-500 bg-purple-500/10',
-        };
-    if (audioExts.includes(ext))
-        return {
-            icon: FileAudio,
-            colorClass: 'text-amber-500 bg-amber-500/10',
-        };
-    if (archiveExts.includes(ext))
-        return {
-            icon: FileArchive,
-            colorClass: 'text-orange-500 bg-orange-500/10',
-        };
-    if (codeExts.includes(ext))
-        return {
-            icon: FileCode,
-            colorClass: 'text-emerald-500 bg-emerald-500/10',
-        };
-    if (docExts.includes(ext))
-        return { icon: FileText, colorClass: 'text-blue-500 bg-blue-500/10' };
-    return { icon: FileIcon, colorClass: 'text-muted-foreground bg-muted' };
-}
-
-function StatusDot({ status }: { status: DerivedStatus }) {
-    return (
-        <span className="inline-flex items-center gap-1.5">
-            <span
-                className={cn(
-                    'relative inline-block size-2 rounded-full',
-                    status === 'archived' && 'bg-muted-foreground/50',
-                    status === 'retrieving' && 'bg-blue-500',
-                    status === 'available' && 'bg-emerald-500'
-                )}
-            >
-                {status === 'retrieving' && (
-                    <span className="absolute inset-0 animate-ping rounded-full bg-blue-500/60" />
-                )}
-            </span>
-            <span
-                className={cn(
-                    'text-xs capitalize',
-                    status === 'archived' && 'text-muted-foreground',
-                    status === 'retrieving' &&
-                        'text-blue-600 dark:text-blue-400',
-                    status === 'available' &&
-                        'text-emerald-600 dark:text-emerald-400'
-                )}
-            >
-                {status}
-            </span>
-        </span>
-    );
-}
-
-/**
- * File icon that morphs into a checkbox on hover or when in selection mode.
- * Renders at a fixed size to prevent layout shift.
- */
-function SelectableIcon({
-    name,
-    checked,
-    onCheckedChange,
-    showCheckbox,
-    size = 'sm',
-}: {
-    name: string;
-    checked: boolean;
-    onCheckedChange: () => void;
-    showCheckbox: boolean;
-    size?: 'sm' | 'md';
-}) {
-    const { icon: TypeIcon, colorClass } = getFileTypeInfo(name);
-    const isSmall = size === 'sm';
-    const containerClass = isSmall ? 'size-8' : 'size-10';
-    const iconClass = isSmall ? 'size-4' : 'size-5';
-    const reveal = checked || showCheckbox;
-
-    return (
-        <button
-            type="button"
-            className={cn(
-                'group/icon relative flex shrink-0 items-center justify-center rounded-lg transition-colors',
-                containerClass,
-                reveal ? 'bg-primary/10' : colorClass
-            )}
-            onClick={(e) => {
-                e.stopPropagation();
-                onCheckedChange();
-            }}
-            aria-label={`Select ${name}`}
-        >
-            <TypeIcon
-                className={cn(
-                    iconClass,
-                    'transition-opacity',
-                    reveal
-                        ? 'opacity-0'
-                        : 'opacity-100 group-hover/icon:opacity-0'
-                )}
-            />
-            <div
-                className={cn(
-                    'absolute inset-0 flex items-center justify-center transition-opacity',
-                    reveal
-                        ? 'opacity-100'
-                        : 'opacity-0 group-hover/icon:opacity-100'
-                )}
-            >
-                <Checkbox
-                    checked={checked}
-                    tabIndex={-1}
-                    onCheckedChange={onCheckedChange}
-                    aria-hidden
-                />
-            </div>
-        </button>
-    );
-}
 
 interface FileBrowserProps {
     /** Deep-link target (`?file={id}`): scroll to and highlight this file. */
@@ -808,438 +582,5 @@ export function FileBrowser({ focusFileId }: FileBrowserProps) {
                 </div>
             )}
         </div>
-    );
-}
-
-function formatBatchDate(date: Date | null): string | null {
-    if (!date) return null;
-    const d = new Date(date);
-    const sameYear = d.getFullYear() === new Date().getFullYear();
-    return d.toLocaleDateString('en-US', {
-        month: 'short',
-        day: 'numeric',
-        year: sameYear ? undefined : 'numeric',
-    });
-}
-
-interface BatchHeaderProps {
-    group: FileBatchGroup;
-    expanded: boolean;
-    onToggle: () => void;
-}
-
-function BatchHeader({ group, expanded, onToggle }: BatchHeaderProps) {
-    const isUngrouped = group.batchId === null;
-    const fileCount = group.files.length;
-    const totalBytes = group.files.reduce((s, f) => s + (f.size ?? 0), 0);
-    const dateLabel = formatBatchDate(group.batchCreatedAt);
-
-    return (
-        <div className="flex w-full items-center gap-2">
-            <button
-                type="button"
-                onClick={onToggle}
-                aria-expanded={expanded}
-                className="group flex min-w-0 flex-1 cursor-pointer items-center gap-3 rounded-md px-2 py-2.5 text-left transition-colors hover:bg-muted/50"
-            >
-                <ChevronRight
-                    className={cn(
-                        'h-4 w-4 shrink-0 text-muted-foreground transition-transform duration-200',
-                        expanded && 'rotate-90'
-                    )}
-                />
-                <div className="min-w-0 flex-1">
-                    <h3
-                        className={cn(
-                            'truncate font-semibold tracking-tight',
-                            isUngrouped
-                                ? 'text-muted-foreground'
-                                : 'text-foreground'
-                        )}
-                    >
-                        {isUngrouped ? 'Ungrouped' : group.batchName}
-                    </h3>
-                    <p className="mt-0.5 text-xs tabular-nums text-muted-foreground">
-                        {fileCount} {fileCount === 1 ? 'file' : 'files'}
-                        {' · '}
-                        {formatBytes(totalBytes)}
-                        {dateLabel && ` · ${dateLabel}`}
-                    </p>
-                </div>
-            </button>
-            {!isUngrouped && group.batchId && (
-                <BatchRestoreSlot batchId={group.batchId} files={group.files} />
-            )}
-        </div>
-    );
-}
-
-function BatchHeaderRow({
-    group,
-    expanded,
-    onToggle,
-    colSpan,
-}: BatchHeaderProps & { colSpan: number }) {
-    return (
-        <TableRow className="border-b-0 hover:bg-transparent">
-            <TableCell colSpan={colSpan} className="p-0">
-                <BatchHeader
-                    group={group}
-                    expanded={expanded}
-                    onToggle={onToggle}
-                />
-            </TableCell>
-        </TableRow>
-    );
-}
-
-interface BatchRestoreSlotProps {
-    batchId: string;
-    files: FileWithRetrieval[];
-}
-
-function BatchRestoreSlot({ batchId, files }: BatchRestoreSlotProps) {
-    const trpc = useTRPC();
-    const invalidateFileList = useInvalidateFileList();
-    const [isDialogOpen, setIsDialogOpen] = useState(false);
-    const mutation = useMutation(
-        trpc.files.requestBatchRetrieval.mutationOptions({
-            onSuccess() {
-                invalidateFileList();
-                toast.success('Batch retrieval submitted');
-            },
-            onError(err) {
-                toast.error(err.message || 'Failed to request batch retrieval');
-            },
-        })
-    );
-
-    const fileCount = files.length;
-    const restoringCount = files.filter(
-        (f) => deriveStatus(f) === 'retrieving'
-    ).length;
-    const eligibleFiles = files.filter((f) => deriveStatus(f) === 'archived');
-
-    if (restoringCount === fileCount && fileCount > 0) {
-        return (
-            <span className="shrink-0 pr-2 text-xs tabular-nums text-muted-foreground">
-                Restoring {restoringCount}/{fileCount}
-            </span>
-        );
-    }
-    if (restoringCount > 0) {
-        return (
-            <span className="flex shrink-0 items-center gap-1.5 pr-2 text-xs tabular-nums text-muted-foreground">
-                <RotateCw className="h-3.5 w-3.5 animate-spin" />
-                Restoring {restoringCount} of {fileCount}
-            </span>
-        );
-    }
-    if (eligibleFiles.length === 0) return null;
-
-    return (
-        <>
-            <Button
-                variant="outline"
-                size="sm"
-                onClick={() => setIsDialogOpen(true)}
-                disabled={mutation.isPending}
-                className="shrink-0"
-            >
-                <RotateCw
-                    className={cn(
-                        'mr-1.5 h-3.5 w-3.5',
-                        mutation.isPending && 'animate-spin'
-                    )}
-                />
-                {mutation.isPending ? 'Requesting…' : 'Restore batch'}
-            </Button>
-            <RetrieveDialog
-                open={isDialogOpen}
-                onOpenChange={setIsDialogOpen}
-                tiers={eligibleFiles.map((f) => f.storageTier)}
-                fileCount={eligibleFiles.length}
-                onConfirm={() => mutation.mutate({ batchId, tier: 'standard' })}
-            />
-        </>
-    );
-}
-
-interface FileItemProps {
-    file: FileWithRetrieval;
-    isSelected: boolean;
-    isHighlighted: boolean;
-    hasSelection: boolean;
-    onSelect: (shiftKey: boolean) => void;
-}
-
-function useFileActions(file: FileWithRetrieval) {
-    const trpc = useTRPC();
-    const queryClient = useQueryClient();
-    const invalidateFileList = useInvalidateFileList();
-
-    const deleteMutation = useMutation(
-        trpc.files.delete.mutationOptions({
-            onSuccess: invalidateFileList,
-        })
-    );
-
-    const retrievalMutation = useMutation(
-        trpc.files.requestRetrieval.mutationOptions({
-            onSuccess() {
-                invalidateFileList();
-                toast.success('Retrieval request submitted');
-            },
-        })
-    );
-
-    async function handleDownload() {
-        try {
-            const { url } = await queryClient.fetchQuery(
-                trpc.files.getDownloadUrl.queryOptions({ fileId: file.id })
-            );
-            window.open(url, '_blank');
-        } catch {
-            toast.error('Failed to get download URL');
-        }
-    }
-
-    return {
-        onDelete: () => deleteMutation.mutate({ id: file.id }),
-        onRetrieval: () => retrievalMutation.mutate({ fileId: file.id }),
-        onDownload: handleDownload,
-        isDeleting: deleteMutation.isPending,
-        isRetrieving: retrievalMutation.isPending,
-    };
-}
-
-function FileRow({
-    file,
-    isSelected,
-    isHighlighted,
-    hasSelection,
-    onSelect,
-}: FileItemProps) {
-    const status = deriveStatus(file);
-    const actions = useFileActions(file);
-    const ext = getFileExtension(file.name);
-    const downloadWindow = getDownloadWindowLabel(file);
-
-    return (
-        <TableRow
-            data-file-id={file.id}
-            data-state={isSelected ? 'selected' : undefined}
-            className={cn(
-                'cursor-pointer transition-colors',
-                isSelected &&
-                    'bg-primary/4 hover:bg-primary/6 dark:bg-primary/8 dark:hover:bg-primary/10',
-                isHighlighted && 'bg-primary/10 dark:bg-primary/15'
-            )}
-            onClick={(e) => onSelect(e.shiftKey)}
-        >
-            <TableCell className="pl-4">
-                <SelectableIcon
-                    name={file.name}
-                    checked={isSelected}
-                    onCheckedChange={() => onSelect(false)}
-                    showCheckbox={hasSelection}
-                    size="sm"
-                />
-            </TableCell>
-            <TableCell>
-                <div className="min-w-0">
-                    <p className="truncate font-medium leading-tight">
-                        {file.name}
-                    </p>
-                    {ext && (
-                        <p className="text-xs uppercase text-muted-foreground">
-                            {ext}
-                        </p>
-                    )}
-                </div>
-            </TableCell>
-            <TableCell className="tabular-nums text-muted-foreground">
-                {formatBytes(file.size)}
-            </TableCell>
-            <TableCell className="text-muted-foreground">
-                {formatDate(file.createdAt)}
-            </TableCell>
-            <TableCell>
-                <StatusDot status={status} />
-                {downloadWindow && (
-                    <p className="mt-0.5 text-xs text-muted-foreground">
-                        {downloadWindow}
-                    </p>
-                )}
-            </TableCell>
-            <TableCell onClick={(e) => e.stopPropagation()}>
-                <FileActions
-                    status={status}
-                    storageTier={file.storageTier}
-                    {...actions}
-                />
-            </TableCell>
-        </TableRow>
-    );
-}
-
-function FileCard({
-    file,
-    isSelected,
-    isHighlighted,
-    hasSelection,
-    onSelect,
-}: FileItemProps) {
-    const status = deriveStatus(file);
-    const actions = useFileActions(file);
-    const ext = getFileExtension(file.name);
-    const downloadWindow = getDownloadWindowLabel(file);
-
-    return (
-        <Card
-            data-file-id={file.id}
-            className={cn(
-                'group relative cursor-pointer py-0 transition-all',
-                isSelected
-                    ? 'ring-2 ring-primary/30 bg-primary/2 dark:bg-primary/6'
-                    : 'hover:border-border/80',
-                isHighlighted && 'ring-2 ring-primary/40'
-            )}
-            onClick={(e) => onSelect(e.shiftKey)}
-        >
-            <CardContent className="p-4">
-                <div className="mb-3 flex items-start justify-between">
-                    <SelectableIcon
-                        name={file.name}
-                        checked={isSelected}
-                        onCheckedChange={() => onSelect(false)}
-                        showCheckbox={hasSelection}
-                        size="md"
-                    />
-                    <div
-                        className={cn(
-                            'transition-opacity',
-                            hasSelection
-                                ? 'opacity-100'
-                                : 'opacity-0 group-hover:opacity-100 [&:has([data-state=open])]:opacity-100'
-                        )}
-                        onClick={(e) => e.stopPropagation()}
-                    >
-                        <FileActions
-                            status={status}
-                            storageTier={file.storageTier}
-                            {...actions}
-                        />
-                    </div>
-                </div>
-                <p className="truncate text-sm/tight font-medium">
-                    {file.name}
-                </p>
-                <div className="mt-1.5 flex items-center justify-between">
-                    <span className="text-xs tabular-nums text-muted-foreground">
-                        {formatBytes(file.size)}
-                        {ext && (
-                            <>
-                                <span className="mx-1 text-border">/</span>
-                                <span className="uppercase">{ext}</span>
-                            </>
-                        )}
-                    </span>
-                    <div className="flex flex-col items-end gap-0.5">
-                        <StatusDot status={status} />
-                        {downloadWindow && (
-                            <p className="text-xs text-muted-foreground">
-                                {downloadWindow}
-                            </p>
-                        )}
-                    </div>
-                </div>
-            </CardContent>
-        </Card>
-    );
-}
-
-interface FileActionsProps {
-    status: DerivedStatus;
-    storageTier: FileWithRetrieval['storageTier'];
-    onDelete: () => void;
-    onRetrieval: () => void;
-    onDownload: () => void;
-    isDeleting: boolean;
-    isRetrieving: boolean;
-}
-
-function FileActions({
-    status,
-    storageTier,
-    onDelete,
-    onRetrieval,
-    onDownload,
-    isDeleting,
-    isRetrieving,
-}: FileActionsProps) {
-    // The dialog lives outside the dropdown: menu content unmounts on close,
-    // which would tear the dialog down mid-open.
-    const [isRetrieveDialogOpen, setIsRetrieveDialogOpen] = useState(false);
-    return (
-        <>
-            <RetrieveDialog
-                open={isRetrieveDialogOpen}
-                onOpenChange={setIsRetrieveDialogOpen}
-                tiers={[storageTier]}
-                fileCount={1}
-                onConfirm={onRetrieval}
-            />
-            <DropdownMenu>
-                <DropdownMenuTrigger
-                    render={<Button variant="ghost" size="icon-sm" />}
-                >
-                    <MoreHorizontal className="size-4" />
-                    <span className="sr-only">Actions</span>
-                </DropdownMenuTrigger>
-                <DropdownMenuPositioner align="end">
-                    <DropdownMenuContent>
-                        {status === 'archived' && (
-                            <DropdownMenuItem
-                                onClick={() => setIsRetrieveDialogOpen(true)}
-                                disabled={isRetrieving}
-                            >
-                                {isRetrieving ? (
-                                    <Loader2 className="mr-2 size-4 animate-spin" />
-                                ) : (
-                                    <Clock className="mr-2 size-4" />
-                                )}
-                                Request retrieval
-                            </DropdownMenuItem>
-                        )}
-                        {status === 'available' && (
-                            <DropdownMenuItem onClick={onDownload}>
-                                <Download className="mr-2 size-4" />
-                                Download
-                            </DropdownMenuItem>
-                        )}
-                        {status === 'retrieving' && (
-                            <DropdownMenuItem disabled>
-                                <RotateCw className="mr-2 size-4" />
-                                Retrieving...
-                            </DropdownMenuItem>
-                        )}
-                        <DropdownMenuSeparator />
-                        <DropdownMenuItem
-                            className="text-destructive focus:text-destructive"
-                            onClick={onDelete}
-                            disabled={isDeleting}
-                        >
-                            {isDeleting ? (
-                                <Loader2 className="mr-2 size-4 animate-spin" />
-                            ) : (
-                                <Trash2 className="mr-2 size-4" />
-                            )}
-                            Delete
-                        </DropdownMenuItem>
-                    </DropdownMenuContent>
-                </DropdownMenuPositioner>
-            </DropdownMenu>
-        </>
     );
 }
