+++ apps/web/components/dashboard/file-browser/index.ts
@@ -1,0 +1,1 @@
+export { FileBrowser } from './FileBrowser';
