+++ apps/web/components/dashboard/file-browser/types.ts
@@ -1,0 +1,9 @@
+import type { FileWithRetrieval } from '@nexus/db/repo/files';
+
+export interface FileItemProps {
+    file: FileWithRetrieval;
+    isSelected: boolean;
+    isHighlighted: boolean;
+    hasSelection: boolean;
+    onSelect: (shiftKey: boolean) => void;
+}
